WIP: check memory safety guarantees
The destroy function can fail in a number of ways. If it just fails
on the canary validation or on making the memory mutable, we can
retry these operations and still be in safe-land. If however it fails
after freeing the memory (this can't happen in the current implementation
but it's worth talking about) then the purge function will have a
use-after-free vulnerability.

Need to audit this code.